chore: add new ng proxy structure
