+++ apps/angular/package.json
@@ -24,7 +24,6 @@
     "@angular/forms": "^16.1.0",
     "@angular/platform-browser": "^16.1.0",
     "@angular/platform-browser-dynamic": "^16.1.0",
-    "@angular/router": "^16.1.0",
     "@baloise/design-system-brand-icons": "^14.4.0",
     "@baloise/design-system-components": "14.4.0",
     "@baloise/design-system-components-angular": "14.4.0",
