+++ apps/angular/src/app/app.config.ts
@@ -1,10 +1,7 @@
 import { ApplicationConfig, importProvidersFrom } from '@angular/core'
-import { provideRouter } from '@angular/router'
 
-import { routes } from './app.routes'
-
-import { BaloiseDesignSystemModule } from '../generated/src'
+import { BaloiseDesignSystemModule } from '@baloise/design-system-components-angular'
 
 export const appConfig: ApplicationConfig = {
-  providers: [provideRouter(routes), importProvidersFrom(BaloiseDesignSystemModule.forRoot())],
+  providers: [importProvidersFrom(BaloiseDesignSystemModule.forRoot())],
 }
