+++ apps/angular/src/app/form-components/time.component.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core'
 import { CommonModule } from '@angular/common'
-import { BaloiseDesignSystemModule } from 'src/generated/src'
+import { BaloiseDesignSystemModule } from '@baloise/design-system-components-angular'
 import { FormGroup, ReactiveFormsModule } from '@angular/forms'
 import { UpdateControl } from '../app.component'
