+++ apps/angular/tsconfig.json
@@ -23,7 +23,11 @@
     "lib": [
       "ES2022",
       "dom"
-    ]
+    ],
+    "paths": {
+      "@baloise/design-system-components-angular/common": ["src/generated/common/src"],
+      "@baloise/design-system-components-angular": ["src/generated/src"]
+    }
   },
   "angularCompilerOptions": {
     "enableI18nLegacyMessageIdFormat": false,
