+++ packages/components-angular/common/src/config.ts
@@ -1,0 +1,10 @@
+import { BalConfig } from '@baloise/design-system-components'
+
+export interface BaloiseDesignSystemAngularConfig {
+  applyPolyfills?: boolean
+  defaults?: BalConfig
+  forms?: {
+    setInvalid?: boolean
+    invalidateOn?: 'touched' | 'dirty'
+  }
+}
