+++ packages/components-angular/common/src/index.ts
@@ -1,0 +1,16 @@
+export type { BaloiseDesignSystemAngularConfig } from './config'
+export type { ProxyComponent } from './utils'
+
+export { AngularDelegate } from './providers/angular-delegate'
+export { BalBreakpointsService } from './providers/breakpoints.service'
+export { BalConfigService } from './providers/config.service'
+export { BalModalService } from './providers/modal.service'
+export { BalOrientationService } from './providers/orientation.service'
+export { BalSnackbarService } from './providers/snackbar.service'
+export { BalToastService } from './providers/toast.service'
+
+export { BalNgErrorComponent } from './directives/error.component'
+export { BalAutoFocus } from './directives/focus.directive'
+
+export { BalConfigToken } from './token'
+export { raf, parseCustomEvent, element } from './utils'
