+++ packages/components-angular/common/src/providers/config.service.ts
