+++ packages/components-angular/common/src/providers/orientation.service.ts
