+++ packages/components-angular/common/src/providers/snackbar.service.ts
