+++ packages/components-angular/common/src/token.ts
@@ -1,0 +1,3 @@
+import { InjectionToken } from '@angular/core'
+
+export const BalConfigToken = new InjectionToken<any>('USERCONFIG')
