+++ packages/components-angular/common/src/utils.ts
@@ -1,3 +1,16 @@
+declare const __zone_symbol__requestAnimationFrame: any
+declare const requestAnimationFrame: any
+
+export const raf = (h: any): any => {
+  if (typeof __zone_symbol__requestAnimationFrame === 'function') {
+    return __zone_symbol__requestAnimationFrame(h)
+  }
+  if (typeof requestAnimationFrame === 'function') {
+    return requestAnimationFrame(h)
+  }
+  return setTimeout(h)
+}
+
 export const parseCustomEvent = <T>(event: CustomEvent<T> | Event): T => {
   if ('detail' in event) {
     return event.detail
