+++ packages/components-angular/src/index.ts
@@ -1,27 +1,26 @@
-// COMPONENTS
-export { BalNgErrorComponent } from './components/error.component'
+export type { ProxyComponent, BaloiseDesignSystemAngularConfig } from '@baloise/design-system-components-angular/common'
 
-// DIRECTIVES
+export {
+  AngularDelegate,
+  BalBreakpointsService,
+  BalConfigService,
+  BalModalService,
+  BalOrientationService,
+  BalSnackbarService,
+  BalToastService,
+  BalConfigToken,
+  raf,
+  parseCustomEvent,
+  element,
+} from '@baloise/design-system-components-angular/common'
+
 export * from './generated/proxies'
 export { BooleanValueAccessor } from './generated/boolean-value-accessor'
 export { NumericValueAccessor } from './generated/number-value-accessor'
 export { SelectValueAccessor } from './generated/select-value-accessor'
 export { TextValueAccessor } from './generated/text-value-accessor'
-export { BalAutoFocus } from './focus.directive'
 
-// SERVICES
-export { AngularDelegate } from './overlays/angular-delegate'
-export { BalToastService } from './overlays/toast.service'
-export { BalSnackbarService } from './overlays/snackbar.service'
-export { BalModalService } from './overlays/modal.service'
-export { BalBreakpointsService } from './util/breakpoints.service'
-export { BalOrientationService } from './util/orientation.service'
-export { BalConfigService } from './util/config.service'
-export { raf } from './util/util'
+export { BalNgErrorComponent } from './directives/error.component'
+export { BalAutoFocus } from './directives/focus.directive'
 
-// PACKAGE MODULE
-export type { BaloiseDesignSystemAngularConfig } from './app-initialize'
-export { BaloiseDesignSystemModule, BalConfigToken } from './module'
-
-// HELPERS
-export { element, parseCustomEvent, ProxyComponent } from './helpers'
+export { BaloiseDesignSystemModule } from './module'
