+++ packages/components-angular/src/module.ts
@@ -1,23 +1,26 @@
-import { APP_INITIALIZER, InjectionToken, ModuleWithProviders, NgModule, NgZone } from '@angular/core'
+import { NgModule, APP_INITIALIZER, ModuleWithProviders, NgZone } from '@angular/core'
 import { CommonModule, DOCUMENT } from '@angular/common'
-
-import { appInitialize, BaloiseDesignSystemAngularConfig } from './app-initialize'
+import {
+  AngularDelegate,
+  BalBreakpointsService,
+  BalConfigService,
+  BalModalService,
+  BalOrientationService,
+  BalSnackbarService,
+  BalToastService,
+  BalConfigToken,
+} from '@baloise/design-system-components-angular/common'
+import { appInitialize } from './app-initialize'
 import { DIRECTIVES } from './generated/proxies-list'
 import { BooleanValueAccessor } from './generated/boolean-value-accessor'
 import { NumericValueAccessor } from './generated/number-value-accessor'
 import { SelectValueAccessor } from './generated/select-value-accessor'
 import { TextValueAccessor } from './generated/text-value-accessor'
-import { BalAutoFocus } from './focus.directive'
-import { AngularDelegate } from './overlays/angular-delegate'
-import { BalModalService } from './overlays/modal.service'
-import { BalToastService } from './overlays/toast.service'
-import { BalSnackbarService } from './overlays/snackbar.service'
-import { BalNgErrorComponent } from './components/error.component'
-import { BalBreakpointsService } from './util/breakpoints.service'
-import { BalOrientationService } from './util/orientation.service'
-import { BalConfigService } from './util/config.service'
 
-export const BalConfigToken = new InjectionToken<any>('USERCONFIG')
+import { BalNgErrorComponent } from './directives/error.component'
+import { BalAutoFocus } from './directives/focus.directive'
+
+import type { BaloiseDesignSystemAngularConfig } from '@baloise/design-system-components-angular/common'
 
 const DECLARATIONS = [
   // generated proxies
