+++ packages/components/package.json
@@ -90,5 +90,6 @@
     "@baloise/design-system-output-target-react": "14.4.0",
     "@baloise/design-system-output-target-vue": "14.4.0"
   },
-  "gitHead": "4157249628694f5811f322e0a547989ca9a9c6a6"
+  "gitHead": "4157249628694f5811f322e0a547989ca9a9c6a6",
+  "sideEffects": false
 }
